+++ jdaviz/app.py
@@ -498,11 +498,25 @@ class Application(VuetifyTemplate, HubListener):
                      LinkSame(linked_data.components[0], ref_data.components[1])]
             dc.add_link(links)
             return
-        elif ((self.config in ('specviz2d', 'mosviz') and
-               linked_data.meta.get('Plugin', None) is not None) or
-              (linked_data.meta.get('Plugin', None) == 'GaussianSmooth' and
-               linked_data.ndim < 3 and  # Cube linking requires special logic. See below
-               ref_data.ndim < 3)
+
+        elif self.config == 'cubeviz' and linked_data.ndim == 1:
+            ref_wavelength_component = dc[0].components[-2]
+            ref_flux_component = dc[0].components[-1]
+            linked_wavelength_component = dc[-1].components[1]
+            linked_flux_component = dc[-1].components[-1]
+
+            links = [
+                LinkSame(ref_wavelength_component, linked_wavelength_component),
+                LinkSame(ref_flux_component, linked_flux_component)
+            ]
+
+            dc.add_link(links)
+            return
+
+        elif (linked_data.meta.get('Plugin', None) == 'SpectralExtraction' or
+                (linked_data.meta.get('Plugin', None) == ('GaussianSmooth') and
+                 linked_data.ndim < 3 and  # Cube linking requires special logic. See below
+                 ref_data.ndim < 3)
               ):
             links = [LinkSame(linked_data.components[0], ref_data.components[0]),
                      LinkSame(linked_data.components[1], ref_data.components[1])]
@@ -2314,9 +2328,9 @@ class Application(VuetifyTemplate, HubListener):
             tray_item_instance = tray.get('cls')(
                 app=self, **optional_tray_kwargs
             )
-            tray_item_label = tray.get('label')
             # store a copy of the tray name in the instance so it can be accessed by the
             # plugin itself
+            tray_item_label = tray.get('label')
             tray_item_instance._plugin_name = tray_item_label
 
             self.state.tray_items.append({