+++ jdaviz/configs/default/plugins/line_lists/line_lists.py
@@ -4,6 +4,7 @@ import os
 import astropy.units as u
 from astropy import constants as const
 from astropy.table import QTable
+
 from glue.core.message import (SubsetCreateMessage,
                                SubsetDeleteMessage,
                                SubsetUpdateMessage)